pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chorus"
version = "0.1.0"
description = "Personality-driven multi-agent AI framework"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2",
    "pyyaml>=6",
    "anthropic>=0.25",
    "openai>=1.30",
    "python-dotenv>=1.0",
    "ollama>=0.3",
]

[project.scripts]
chorus = "chorus.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["chorus*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
